Huge refactoring commit
All of the functions that can be pulled in from any of SquirrelChat's
headers are now prefixed with "sqchat_", along with any global variables
that can be made visible with the headers. In addition, some variable
names that really just didn't make much sense have been changed to
clarify their meanings.